requirements.txt
Flask==2.2.2
streamlit==1.18.1
requests==2.28.1
textblob==0.15.3
elevenlabs==0.0.1  
python-dotenv==0.19.2
